Move some damage functions to a central place
Move from MeterBrowseWnd and ValueRefs::TotalFighterShots to CombatDamage
