+++ combat/CMakeLists.txt
@@ -1,10 +1,12 @@
 target_sources(freeorioncommon
     PUBLIC
+        ${CMAKE_CURRENT_LIST_DIR}/CombatDamage.h
         ${CMAKE_CURRENT_LIST_DIR}/CombatEvent.h
         ${CMAKE_CURRENT_LIST_DIR}/CombatEvents.h
         ${CMAKE_CURRENT_LIST_DIR}/CombatLogManager.h
         ${CMAKE_CURRENT_LIST_DIR}/CombatSystem.h
     PRIVATE
+        ${CMAKE_CURRENT_LIST_DIR}/CombatDamage.cpp
         ${CMAKE_CURRENT_LIST_DIR}/CombatEvent.cpp
         ${CMAKE_CURRENT_LIST_DIR}/CombatEvents.cpp
         ${CMAKE_CURRENT_LIST_DIR}/CombatLogManager.cpp
