+++ msvc2019/Common/Common.vcxproj.filters
@@ -244,6 +244,9 @@
     <ClInclude Include="..\..\util\ScopedTimer.h">
       <Filter>Header Files\util</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\combat\CombatDamage.h">
+      <Filter>Header Files\combat</Filter>
+    </ClInclude>
     <ClInclude Include="..\..\combat\CombatEvent.h">
       <Filter>Header Files\combat</Filter>
     </ClInclude>
@@ -501,6 +504,9 @@
     </ClCompile>
     <ClCompile Include="..\..\util\SaveGamePreviewUtils.cpp">
       <Filter>Source Files\util</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\combat\CombatDamage.cpp">
+      <Filter>Source Files\combat</Filter>
     </ClCompile>
     <ClCompile Include="..\..\combat\CombatEvents.cpp">
       <Filter>Source Files\combat</Filter>
