+++ universe/ValueRefs.cpp
@@ -26,6 +26,7 @@
 #include "UniverseObjectVisitors.h"
 #include "UniverseObject.h"
 #include "Universe.h"
+#include "../combat/CombatDamage.h"
 #include "../Empire/Empire.h"
 #include "../Empire/Supply.h"
 #include "../util/GameRules.h"
@@ -1395,48 +1396,19 @@ unsigned int TotalFighterShots::GetCheckSum() const
 
     CheckSums::CheckSumCombine(retval, "ValueRef::TotalFighterShots");
     CheckSums::CheckSumCombine(retval, m_sampling_condition);
-    TraceLogger() << "GetCheckSum(Statisic<T>): " << typeid(*this).name() << " retval: " << retval;
+    TraceLogger() << "GetCheckSum(TotalFighterShots):  retval: " << retval;
     return retval;
 }
 
 int TotalFighterShots::Eval(const ScriptingContext& context) const
 {
-    Condition::ObjectSet condition_matches;
-
-    // Iterate over context, but change bout number
-    // XXX probably should rather take ship ID as argument as well
-    // XXX else i get a fighter and have to fetch me the launching ship from that
     std::shared_ptr<const Ship> ship = std::static_pointer_cast<const Ship>(context.source);
     if (!ship) {
         ErrorLogger() << "TotalFighterShots condition used in context where the Source is not a ship";
         return 0;
     }
 
-    //XXX actually part of combat system, defined in AutoResolveCombat, maybe reuse
-    ScriptingContext mut_context(context);
-    int launch_capacity = ship->SumCurrentPartMeterValuesForPartClass(MeterType::METER_CAPACITY, ShipPartClass::PC_FIGHTER_BAY);
-    int hangar_fighters = ship->SumCurrentPartMeterValuesForPartClass(MeterType::METER_CAPACITY, ShipPartClass::PC_FIGHTER_HANGAR);
-    int launched_fighters = 0;
-    int shots_total = 0;
-
-    for (int bout = 1; bout <= GetGameRules().Get<int>("RULE_NUM_COMBAT_ROUNDS"); ++bout) {
-        mut_context.combat_bout = bout;
-        int launch_this_bout = std::min(launch_capacity,hangar_fighters);
-        int shots_this_bout = launched_fighters;
-        if (m_sampling_condition && launched_fighters > 0) {
-            // check if not shooting
-            condition_matches.clear();
-            m_sampling_condition->Eval(mut_context, condition_matches);
-            if (condition_matches.size() == 0) {
-                shots_this_bout = 0;
-            }
-        }
-        shots_total += shots_this_bout;
-        launched_fighters += launch_this_bout;
-        hangar_fighters -= launch_this_bout;
-    }
-
-    return shots_total;
+    return Combat::TotalFighterShots(context, *ship, m_sampling_condition.get());
 }
 
 ///////////////////////////////////////////////////////////
